entregar esa monda

+++ Servidor.py
@@ -48,11 +48,25 @@ class Server():
         self.dicc['mensaje'] =f"el bucket {datos['bucket']} ha sido eliminado"
         self.enviar_archivo()
     
+    def enviar_descargable(self,datos):
+        if(os.path.isfile('Buckets/' + datos['bucket'] +'/'+ datos['nombreArchivo']) == False):
+            print('Buckets/' + datos['bucket'] + datos['nombreArchivo'])
+            self.dicc['mensaje'] =f"El archivo {datos['nombreArchivo']} no existe"
+        else:
+            file = open('Buckets/' + datos['bucket'] +'/'+ datos['nombreArchivo'],'rb')
+            self.dicc['mensaje'] = f"Descarga realiza del archivo {datos['nombreArchivo']}"
+            self.dicc['archivo'] = file.read()
+            self.dicc['nombreArchivo'] = os.path.split('Buckets/' + datos['bucket'] + datos['nombreArchivo'])[1]
+            file.close
+        self.enviar_archivo()
 
     def eliminar_archivo(self,datos):
-        nombre = 'Buckets/' + datos['bucket']+'/' + datos['nombreArchivo']
-        os.remove(nombre)
-        self.dicc['mensaje'] =f"el archivo {datos['nombreArchivo']} ha sido eliminado"
+        if(os.path.isfile('Buckets/' + datos['bucket'] +'/'+ datos['nombreArchivo'])== False):
+            self.dicc['mensaje'] =f"El archivo {datos['nombreArchivo']} no existe"
+        else:
+            nombre = 'Buckets/' + datos['bucket']+'/' + datos['nombreArchivo']
+            os.remove(nombre)
+            self.dicc['mensaje'] =f"el archivo {datos['nombreArchivo']} ha sido eliminado"
         self.enviar_archivo()
     
 
@@ -106,6 +120,7 @@ class Server():
         # swicth para llamara los metodos necesarios que envia el cliente
         self.dicc['lista'] = os.listdir('Buckets')
         self.dicc['mensaje'] = "Esta es la lista de buckets"
+        self.dicc['archivo'] = "no"
         datos = pickle.loads(x)
         if(datos['comando'] == '1'):
             self.crear_bucket(datos)
@@ -126,6 +141,10 @@ class Server():
             self.listar_archivos(datos)
 
         if(datos['comando'] == '7'):
+            print("7")
+            self.enviar_descargable(datos)
+
+        if(datos['comando'] == '8'):
             print("Se esta cerrando el servidor.")
             self.connected = False
